requirements.txt
statsd
pyyaml
mysql-connector-python
cassandra-driver
pymssql